misp/requirements.txt
redis>=5.0,<6.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
